# api/docs/swagger.yaml
basePath: /
definitions:
  api_handlers.CreateMovieRequest:
    properties:
      title:
        example: Interestelar
        type: string
      year:
        example: 2014
        type: integer
    required:
    - title
    - year
    type: object
  github_com_jamescookdev_projeto-sipub-tech_movies-service_gen_go.Movie:
    properties:
      id:
        type: string
      title:
        type: string
      year:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: Esta é uma API REST para consulta e gerenciamento de filmes.
  title: API de Filmes - Microsserviços com Go e gRPC
  version: "1.0"
paths:
  /movies:
    get:
      description: Retorna uma lista de filmes, com a possibilidade de usar limit
        e offset para paginação.
      parameters:
      - default: 20
        description: Número de resultados por página
        in: query
        name: limit
        type: integer
      - default: 0
        description: Número de resultados a pular
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_jamescookdev_projeto-sipub-tech_movies-service_gen_go.Movie'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              allOf:
              - type: string
              - properties:
                  error:
                    type: string
                type: object
            type: object
      summary: Lista os filmes com paginação
      tags:
      - Movies
    post:
      consumes:
      - application/json
      description: Adiciona um novo filme ao banco de dados.
      parameters:
      - description: Dados para criar o filme
        in: body
        name: movie
        required: true
        schema:
          $ref: '#/definitions/api_handlers.CreateMovieRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_jamescookdev_projeto-sipub-tech_movies-service_gen_go.Movie'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              allOf:
              - type: string
              - properties:
                  error:
                    type: string
                type: object
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              allOf:
              - type: string
              - properties:
                  error:
                    type: string
                type: object
            type: object
      summary: Cria um novo filme
      tags:
      - Movies
  /movies/{id}:
    delete:
      description: Remove um filme do banco de dados baseado no seu ID.
      parameters:
      - description: ID do Filme
        format: mongodb-id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              allOf:
              - type: string
              - properties:
                  message:
                    type: string
                type: object
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              allOf:
              - type: string
              - properties:
                  error:
                    type: string
                type: object
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              allOf:
              - type: string
              - properties:
                  error:
                    type: string
                type: object
            type: object
      summary: Deleta um filme
      tags:
      - Movies
    get:
      description: Retorna os detalhes de um filme específico baseado no seu ID.
      parameters:
      - description: ID do Filme
        format: mongodb-id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_jamescookdev_projeto-sipub-tech_movies-service_gen_go.Movie'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              allOf:
              - type: string
              - properties:
                  error:
                    type: string
                type: object
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              allOf:
              - type: string
              - properties:
                  error:
                    type: string
                type: object
            type: object
      summary: Busca um filme por ID
      tags:
      - Movies
swagger: "2.0"